+++ init.go
@@ -32,6 +32,8 @@ func init() {
 	config.Credentials.Username = os.Getenv("AUTH_USERNAME")
 	config.Credentials.Password = os.Getenv("AUTH_PASSWORD")
 	config.AppName = "Mesos Kafka Framework"
+	config.ZookeeperCustomString = os.Getenv("ZOOKEEPER_CUSTOM_DOMAIN")
+	config.KafkaCustomString = os.Getenv("KAFKA_CUSTOM_DOMAIN")
 
 	// The comunication to the mesos server should be via ssl or not
 	if os.Getenv("MESOS_SSL") == "true" {