+++ changelog.md
@@ -15,3 +15,5 @@
 - Add RestCall to kill a task
 - Add Authentication
 - Add Support to configure (non)SSL Support to Mesos
+- Add Custom Domain to Zookeeper and Kafka to match Consul DNS
+- Add Service Name ENV Variable to Zookeeper and Kafka to match Consul DNS